Performance improvements

+++ SampleProject/Motis/Performance Test/MJCompany.m
@@ -44,7 +44,7 @@
 
 - (NSString*)description
 {
-    return [NSString stringWithFormat:@"[%@ - identifier:%d - isActive:%d - yearlyIncome:%f - previousYearIncome:%f - name:%@]", [super description], _identifier, _isActive, _yearlyIncome, _previousYearIncome, _name];
+    return [NSString stringWithFormat:@"[%@ - identifier:%ld - isActive:%ld - yearlyIncome:%f - previousYearIncome:%f - name:%@]", [super description], (long)_identifier, (long)_isActive, _yearlyIncome, _previousYearIncome, _name];
 }
 
 @end
